# docs/specs/001-sakuraplayer-v1/contracts/rest-api.openapi.yaml
openapi: 3.1.0
info:
  title: SakuraPlayer API
  version: 1.2.0
  description: >-
    Private single-administrator API for the SakuraPlayer v1 backend. Video bytes
    are never proxied by this API; signed stream endpoints return 302 redirects.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /settings:
    get:
      tags: [Admin]
      operationId: getSettings
      responses:
        '200':
          description: Redacted settings and connection states.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
    patch:
      tags: [Admin]
      operationId: updateSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsPatch'
      responses:
        '200':
          description: Redacted updated settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '422':
          $ref: '#/components/responses/ValidationError'
  /settings/connection-tests:
    post:
      tags: [Admin]
      operationId: testConnection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target]
              properties:
                target:
                  type: string
                  enum: [cloud115, javdb, dmm, gfriends, ai]
              additionalProperties: false
      responses:
        '200':
          description: Redacted test result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionTest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    ValidationError:
      description: Request cannot be validated.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    ApiError:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          pattern: '^[a-z0-9_]+$'
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        request_id:
          type: string
...
    ProviderState:
      type: object
      required: [configured, status]
      properties:
        configured:
          type: boolean
        status:
          type: string
          enum: [unknown, available, unavailable, credentials_invalid]
        last_checked_at:
          type: [string, 'null']
          format: date-time
        last_error_code:
          type: [string, 'null']
    JavdbSettings:
      allOf:
        - $ref: '#/components/schemas/ProviderState'
        - type: object
          required: [password_configured, version]
          properties:
            username:
              type: [string, 'null']
            password_configured:
              type: boolean
            version:
              type: integer
              minimum: 0
    AiSettings:
      allOf:
        - $ref: '#/components/schemas/ProviderState'
        - type: object
          required: [api_key_configured, version]
          properties:
            base_url:
              type: [string, 'null']
              format: uri
            model:
              type: [string, 'null']
            timeout_seconds:
              type: [integer, 'null']
              minimum: 1
              maximum: 600
            api_key_configured:
              type: boolean
            version:
              type: integer
              minimum: 0
    SyncRunState:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [never, running, succeeded, failed]
        release_id:
          type: [string, 'null']
        started_at:
          type: [string, 'null']
          format: date-time
        completed_at:
          type: [string, 'null']
          format: date-time
        last_successful_at:
          type: [string, 'null']
          format: date-time
        next_scheduled_at:
          type: [string, 'null']
          format: date-time
        last_error_code:
          type: [string, 'null']
    AvdbSyncStatus:
      type: object
      required: [incremental_30d, full_reconcile]
      properties:
        incremental_30d:
          $ref: '#/components/schemas/SyncRunState'
        full_reconcile:
          $ref: '#/components/schemas/SyncRunState'
    Settings:
      type: object
      required:
        [cache_ttl_hours, ready_cache_limit, metadata_concurrency, metadata_timeout_seconds, javdb, ai, providers, avdb_sync]
      properties:
        cache_ttl_hours:
          type: integer
          minimum: 1
          maximum: 168
          default: 24
        ready_cache_limit:
          type: integer
          default: 20
        metadata_concurrency:
          type: integer
          const: 3
        metadata_timeout_seconds:
          type: integer
          const: 600
        javdb:
          $ref: '#/components/schemas/JavdbSettings'
        ai:
          $ref: '#/components/schemas/AiSettings'
        providers:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ProviderState'
        avdb_sync:
          $ref: '#/components/schemas/AvdbSyncStatus'
    SettingsPatch:
      type: object
      minProperties: 1
      properties:
        cache_ttl_hours:
          type: integer
          minimum: 1
          maximum: 168
        javdb:
          $ref: '#/components/schemas/JavdbSettingsCommand'
        ai:
          $ref: '#/components/schemas/AiSettingsCommand'
      additionalProperties: false
    SettingClearCommand:
      type: object
      required: [action, expected_version]
      properties:
        action:
          const: clear
        expected_version:
          type: integer
          minimum: 1
      additionalProperties: false
    JavdbSettingsReplace:
      type: object
      required: [action, expected_version, username, password]
      properties:
        action:
          const: replace
        expected_version:
          type: integer
          minimum: 0
        username:
          type: string
          minLength: 1
          maxLength: 255
        password:
          type: string
          minLength: 1
          maxLength: 1024
      additionalProperties: false
    JavdbSettingsCommand:
      oneOf:
        - $ref: '#/components/schemas/JavdbSettingsReplace'
        - $ref: '#/components/schemas/SettingClearCommand'
    AiSettingsReplace:
      type: object
      required: [action, expected_version, base_url, api_key, model, timeout_seconds]
      properties:
        action:
          const: replace
        expected_version:
          type: integer
          minimum: 0
        base_url:
          type: string
          format: uri
          maxLength: 2048
        api_key:
          type: string
          minLength: 1
          maxLength: 8192
        model:
          type: string
          minLength: 1
          maxLength: 255
        timeout_seconds:
          type: integer
          minimum: 1
          maximum: 600
      additionalProperties: false
    AiSettingsCommand:
      oneOf:
        - $ref: '#/components/schemas/AiSettingsReplace'
        - $ref: '#/components/schemas/SettingClearCommand'
    ConnectionTest:
      type: object
      required: [target, status, error_code, elapsed_ms, checked_at]
      properties:
        target:
          type: string
          enum: [cloud115, javdb, dmm, gfriends, ai]
        status:
          type: string
          enum: [available, unavailable, credentials_invalid, not_configured]
        error_code:
          type: [string, 'null']
        elapsed_ms:
          type: integer
        checked_at:
          type: string
          format: date-time
...